More changes to Product Stats

+++ src/gui/reports/Cell.java
@@ -43,8 +43,8 @@ public class Cell {
 	@Override
 	public String toString(){
 		String result = "";
-		if(value instanceof Long) {
-			result = Long.toString((long)value);
+		if(value instanceof Float) {
+			result = Float.toString((float)value);
 		} else if(value instanceof Integer) {
 			result = Integer.toString((int)value);
 		} else if(value instanceof String) {
